Sets node-red working directory as userData
"--userDir" option sets the directory for projects and flows.
This change sets the node-red working directory as the user userData
so that when projects are cloned, the file input nodes can
navigate to the required supporting files.

See note on file input node:
"The filename should be an absolute path, otherwise it will be relative
 to the working directory of the Node-RED process.""

absolute paths are not desirable