Adds dependencies for harvest-flows
This is a temporary arrangement.
I'm explicitly adding the dependencies listed in share-red-flows
https://github.com/share-research/share-red-flows/blob/master/package.json#L5

This is non-ideal because different flows will have different dependencies.
So we don't want to install them if they're not needed by a flow.

Here's a intial thought dump on how this could be handled:
* Manage multiple prebuilt docker images for each flow. If there are a
lot of flows then this approach will need process rigor, and can become
unmanageable pretty quickly.
* Build a new image on the fly at container deploy time based on recent
changes to the flow deps. This approach will add some extra steps at
deploy time but is a more reliable "as-needed" approach.